requirements.txt
coord_convert>=0.2.1
flasgger>=0.9.7.1
Flask>=2.2.2
flask_restx>=1.2.0
intersect>=1.2
joblib>=1.2.0
numpy>=1.24.3
pandas>=1.4
plotly>=5.18.0
pyproj>=3.6.0
scikit_learn>=1.3.0
scipy>=1.10.1
Shapely>=2.0.3
waitress>=2.0.0
Cython>=3.0.9
pyyaml>=6.0
cryptography>=42.0.3
distro>=1.9.0
pymysql>=1.1.1
pywavelets>=1.8.0
ewtpy>=0.2
matplotlib>=3.9.0